yze blizko

+++ src/components/changeForm/change.jsx
@@ -3,6 +3,7 @@ import ChangeFields from "./changeFields";
 import axios from "axios";
 import {useState} from "react";
 
+
 function ChangeForm() {
 
     const[field, setField] = useState('')
@@ -25,10 +26,11 @@ function ChangeForm() {
         )}
 
         let handleDelete = async () => {
-            let res = await axios.delete(`http://uni-team-inc.online:8080/api/v1/delete?form=${uuid}`)
-        .then(res => {
-            console.log(res.data)
-        })}
+            let result = window.confirm("Вы уверенны?");
+                if (result === true){
+                    let res = await axios.delete(`http://uni-team-inc.online:8080/api/v1/delete?form=${uuid}`)
+                }
+    }
 
     if (data.Fields !== null) {
         return (
@@ -54,15 +56,23 @@ function ChangeForm() {
         )}
 
     return (
-        <div className="changeContainer">
-            <div className="changeContainerContent">
-                <h1>{data.Name}</h1>
-                <p> {data.Description}</p><br/>
-                <div className="wrapperBtn">
-                    <input type="text" placeholder="Добавить поле" maxLength="50" value={field} onChange={event => setField(event.target.value)}/>
-                    <button className="btn" onClick={(e) => {handleSubmit(e)}}>+</button>
-                </div>
-            </div><br/>
+        <div className="mainChangeContainer">
+            <div className="btnDelContainer">
+                <label htmlFor="">
+                    Удалить форму
+                    <button className="btnDel" onClick={(e) => {handleDelete(e)}}>-</button>
+                </label>
+            </div>
+            <div className="changeContainer">
+                <div className="changeContainerContent">
+                    <h1>{data.Name}</h1>
+                    <p> {data.Description}</p><br/>
+                    <div className="wrapperBtn">
+                        <input type="text" placeholder="Добавить поле" maxLength="50" value={field} onChange={event => setField(event.target.value)}/>
+                        <button className="btn" onClick={(e) => {handleSubmit(e)}}>+</button>
+                    </div>
+                </div><br/>
+            </div>
         </div>
     )}
